refactor: separate tracker under source studies maincolumn and add sourcestudies wrapper (#2936)
