--- components/HCABioNetworks/Network/Atlas/components/SourceStudies/components/MainColumn/Table/tracker/types.ts
@@ -1,5 +1,0 @@
-import type { TrackerSourceStudy } from "../../../../../../../../../../@types/network";
-
-export interface Props {
-  data: TrackerSourceStudy[];
-}
