+++ components/HCABioNetworks/Network/Atlas/components/SourceStudies/components/MainColumn/tracker/components/Table/types.ts
@@ -1,0 +1,5 @@
+import type { TrackerSourceStudy } from "../../../../../../../../../../../@types/network";
+
+export interface Props {
+  data: TrackerSourceStudy[];
+}
