+++ components/HCABioNetworks/Network/Atlas/components/SourceStudies/components/MainColumn/tracker/sourceStudies.tsx
@@ -1,25 +1,26 @@
 import { FluidPaper } from "@databiosphere/findable-ui/lib/components/common/Paper/components/FluidPaper/fluidPaper";
 import { GridPaper } from "@databiosphere/findable-ui/lib/components/common/Paper/paper.styles";
 import { JSX } from "react";
-import { Table as CommonTable } from "../../../../../../../../../common/Table/table";
-import { useTable } from "./hook";
-import type { Props } from "./types";
+import { Table } from "../../../../../../../../common/Table/table";
+import { useTable } from "./components/Table/hook";
+import { Props } from "./types";
 
 /**
- * Tracker source studies table for the Source Studies tab.
+ * Tracker source studies view for the Source Studies tab.
  * @param props - Component props.
  * @param props.data - Tracker source studies.
- * @returns table of tracker source studies, or empty state if no data.
+ * @returns tracker source studies view.
  */
-export const Table = ({ data }: Props): JSX.Element => {
+export const SourceStudies = ({ data }: Props): JSX.Element => {
   const table = useTable(data);
+
   return (
     <FluidPaper elevation={0}>
       <GridPaper>
         {table.getRowCount() === 0 ? (
           <div>No Source Studies</div>
         ) : (
-          <CommonTable table={table} />
+          <Table table={table} />
         )}
       </GridPaper>
     </FluidPaper>
